mid is double, check cast
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/segannot@34 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ python/SegAnnot.h
@@ -3,7 +3,7 @@ int SegAnnotBases(
     const int *, const int *,  //annotations
     const int, const int, //sizes
     int *, int*, double*, //segments
-    int*, int*, int*); //breakpoints
+    int*, double*, int*); //breakpoints
 
 #define ERROR_BASES_NOT_INCREASING 1
 #define ERROR_REGIONS_NOT_INCREASING 2
